refactor: move group pages into features with stubs
