+++ src/components/GroupDetailPage.tsx
@@ -1,4 +1,3 @@
-// GroupDetailPage - Re-exporta el Container para mantener compatibilidad de imports
-// Este archivo sirve como punto de entrada para el patrón Container/Presentational
-
-export { GroupDetailContainer as default } from './GroupDetailContainer';
+// TODO: STUB REMOVE BY: 2026-03-31
+// Ticket: PHASE2-007
+export { default } from '@/features/groups/pages/GroupDetailPage';
